Fixing issues with linker files
DNA-BACon connects one end of a linker to an oxygen atom in the 5' phosphate of the first base in the pillar, so the connecting atom in the linker should not be O (meaning that alcohol or phosphate groups might need to be removed). On the other pillar, the 3' oxygen of the last base connects to whatever atom is specified in the linker. A good choice might be PO3H, where the H is bonded directly to P. The O in the pillar can then be attached to P and the program will naturally remove the H, resulting in a normal phosphate connecting the rest of the linker to the DNA. In this edit, I fixed the predefined linker PDB files and their descriptions in the linkerInformation.toml configuration file to produce more reasonable connection behavior. 

However, DNA-BACon still does not check the chirality of the last linker atom after being connected to the pillar